meeting-mate-app/server/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
firebase-admin
pydantic
